Clean up album tile component
